backend/requirements.txt
fastapi
uvicorn
httpx
sqlalchemy
psycopg2-binary
pgvector
llama-index
llama-index-vector-stores-postgres
llama-index-embeddings-ollama
llama-index-llms-ollama
sentence-transformers
python-dotenv
slowapi
sse-starlette
requests
beautifulsoup4
alembic
yfinance
